+++ packages/vda-web3-client/src/config.ts
@@ -3,12 +3,15 @@ import { BigNumber } from '@ethersproject/bignumber'
 import { Contract } from '@ethersproject/contracts'
 import { JsonRpcProvider, Provider } from '@ethersproject/providers'
 import { knownNetworks } from './constants'
-import { Web3ContractInfo, Web3ProviderConfiguration } from '@verida/types'
+import { Web3ContractInfo, Web3ProviderConfiguration, Web3GasConfiguration } from '@verida/types'
 
 export function isVeridaWeb3GasConfiguration(obj : Object) {
     return ('maxFeePerGas' in obj)
         || ('maxPriorityFeePerGas' in obj)
         || ('gasLimit' in obj)
+        || ('gasPrice' in obj)
+        || ('eip1559Mode' in obj)
+        || ('eip1559gasStationUrl' in obj);
 }